move to separate file and disable in race mode as histogram metrics are not enabled
